Code up a model to validate whether logistic regression is right
Ran into some problems. One is that the elo model does not do very well,
no matter how you tweak it. I'm concerned I'm not doing it correctly.
Secondly, it seems that logistic regression is NOT the correct model. I
am going to sleep on this and think for a while.